==> 1-3.ipynb <==
import gdown
gdown.download('https://bit.ly/3eeMKZ', '남산도서관 장서 대출목록 (2023년 04월).csv', quiet=False)

import chardet
with open('남산도서관 장서 대출목록 (2023년 04월).csv', mode='rb') as f:
    d = f.readline()
print(chardet.detect(d))

with open('남산도서관 장서 대출목록 (2023년 04월).csv', encoding='EUC-KR') as f:
    print(f.readline())
    print(f.readline())

import pandas as pd
df = pd.read_csv('남산도서관 장서 대출목록 (2023년 04월).csv', encoding='EUC-KR')

df = pd.read_csv('남산도서관 장서 대출목록 (2023년 04월).csv', encoding='EUC-KR', low_memory=False)

df.head()

df.to_csv('ns_202304.csv')

with open('ns_202304.csv',encoding='utf-8') as f:
    for i in range(3):
        print(f.readline(), end='')

ns_df = pd.read_csv('ns_202304.csv', low_memory=False)
ns_df.head()

ns_df = pd.read_csv('ns_202304.csv', index_col=0, low_memory=False)
ns_df.head()

df.to_csv('ns_202304.csv', index=False)
