==> 3-1.ipynb <==
# 열 삭제하기

import gdown 
gdown.download('https://bit.ly/3RhoNho','ns_202304.csv', quiet=False)

import pandas as pd
ns_df = pd.read_csv('ns_202304.csv',low_memory=False)
ns_df.head()

ns_book = ns_df.loc[:,'번호':'등록일자']
ns_book.head()

print(ns_df.columns)

print(ns_df.columns[0])

ns_df.columns != 'Unnamed: 13'

selected_columns = ns_df.columns != 'Unnamed: 13'
ns_book = ns_df.loc[:,selected_columns]
ns_book.head()

selected_columns = ns_df.columns != '부가기호'
ns_book = ns_df.loc[:,selected_columns]
ns_book.head()

ns_book = ns_df.drop('Unnamed: 13', axis=1)
ns_book.head()

ns_book = ns_df.drop(['부가기호','Unnamed: 13'], axis=1)
ns_book.head()

ns_book.drop('주제분류번호', axis=1, inplace=True)
ns_book.head()

ns_book = ns_df.dropna(axis=1)
ns_book.head()

ns_book = ns_df.dropna(axis=1, how='all')
ns_book.head()

# 행 삭제하기

ns_book2 = ns_book.drop([0,1])
ns_book2.head()

ns_book2 = ns_book[2:]
ns_book2.head()

ns_book2 = ns_book[0:2]
ns_book2.head()

selected_rows = ns_df['출판사'] == '한빛미디어'
ns_book2 = ns_book[selected_rows]
ns_book2.head()

ns_book2 = ns_book[ns_book['대출건수']> 1000]
ns_book.head()

# 중복된 행 찾기

sum(ns_book.duplicated())

sum(ns_book.duplicated(subset=['도서명','저자','ISBN']))

dup_rows=ns_book.duplicated(subset=['도서명','저자','ISBN'], keep=False)
ns_book3 = ns_book[dup_rows]
ns_book3.head()

# 그룹별로 모으기

count_df = ns_book[['도서명','저자','ISBN','권','대출건수']]

group_df=count_df.groupby(by=['도서명','저자','ISBN','권'], dropna=False)
loan_count = group_df.sum()

loan_count = count_df.groupby(by=['도서명','저자','ISBN','권'], dropna=False).sum()
loan_count.head()

# 원본 데이터 업데이트하기

dup_rows = ns_book.duplicated(subset=['도서명','저자','ISBN','권'])

unique_rows = ~dup_rows
ns_book3 = ns_book[unique_rows].copy()

sum(ns_book3.duplicated(subset=['도서명','저자','ISBN','권']))

ns_book3.set_index(['도서명','저자','ISBN','권'], inplace=True)
ns_book3.head()

ns_book3.update(loan_count)
ns_book3.head()

ns_book4 = ns_book3.reset_index()
ns_book4.head()

sum(ns_book['대출건수']>100)

sum(ns_book4['대출건수']>100)

ns_book4=ns_book4[ns_book.columns]
ns_book4.head()

ns_book4.to_csv('ns_book4.csv',index=False)

# 일괄 처리 함수 만들기

def data_cleaning(filename):
    ns_df = pd.read_csv(filename, low_memory=False)
    
    ns_book = ns_df.dropna(axis=1, how='all')
    count_df = ns_book[['도서명','저자','ISBN','권','대출건수']]
    
    loan_count = count_df.groupby(by=['도서명','저자','ISBN','권'], dropna=False). sum()
    
    dup_rows = ns_book.duplicated(subset=['도서명','저자','ISBN','권'])
    unique_rows = ~dup_rows
    ns_book3 = ns_book[unique_rows].copy()
    
    ns_book3.set_index(['도서명','저자','ISBN','권'], inplace=True)
    
    ns_book3.update(loan_count)
    
    ns_book4 = ns_book3.reset_index()
    
    ns_book4 = ns_book4[ns_book.columns]
    
    return ns_book4

new_ns_book4 = data_cleaning('ns_202304.csv')
ns_book4.equals(new_ns_book4)
