==> 3-2.ipynb <==
## 데이터프레임 정보 요약하기

import gdown
gdown.download('https://bit.ly/3GisL6J', 'ns_book4.csv', quiet=False)

import pandas as pd
ns_book4 = pd.read_csv('ns_book4.csv',low_memory=False)
ns_book4.head()

ns_book4.info()

# 누락된 값 처리하기

ns_book4.isna().sum()

ns_book4.loc[0, '도서권수'] = None
ns_book4['도서권수'].isna().sum()

ns_book4.head(2)

ns_book4.loc[0, '도서권수'] = 1
ns_book4 = ns_book4.astype({'도서권수':'int32', '대출건수':'int32'})
ns_book4.head(2)

ns_book4.loc[0, '부가기호'] = None
ns_book4.head(2)

import numpy as np
ns_book4.loc[0, '부가기호'] = np.nan
ns_book4.head(2)

set_isbn_na_rows = ns_book4['세트 ISBN'].isna()

ns_book4.loc[set_isbn_na_rows, '세트 ISBN'] = ''
ns_book4['세트 ISBN'].isna().sum()

ns_book4.fillna('없음').isna().sum()

ns_book4['부가기호'].fillna('없음').isna().sum()

ns_book4.fillna({'부가기호':'없음'}).isna().sum()

ns_book4.replace(np.nan, '없음').isna().sum()

ns_book4.replace([np.nan, '2021'], ['없음', '21']).head(2)

ns_book4.replace({np.nan:'없음', '2021':'21'}).head(2)

ns_book4.replace({'부가기호':np.nan}, '없음').head(2)

ns_book4.replace({'부가기호':{np.nan: '없음'}, '발행년도':{'2021':'21'}}).head(2)

# 정규 표현식

ns_book4.replace({'발행년도':{'2021':'21'}})[100:102]

ns_book4.replace({'발행년도':{r'\d\d(\d\d)':r'\1'}}, regex=True)[100:102]

ns_book4.replace({'발행년도':{r'\d{2}(\d{2})':r'\1'}}, regex=True)[100:102]

ns_book4.replace({'저자':{r'(.*)\s\(지은이\)(.*)\s\(옮긴이\)': r'\1\2'}, '발행년도': {r'\d{2}(\d{2})':r'\1'}}, regex=True)[100:102]

# 잘못된 값 바꾸기

ns_book4.astype({'발행년도':'int32'})

ns_book4['발행년도'].str.contains('1988').sum()

invalid_number = ns_book4['발행년도'].str.contains('\D', na=True)
print(invalid_number.sum())
ns_book4[invalid_number].head()

ns_book5 = ns_book4.replace({'발행년도':r'.*(\d{4}).*'}, r'\1', regex=True)
ns_book5[invalid_number].head()

unkown_year = ns_book5['발행년도'].str.contains('\D', na=True)
print(unkown_year.sum())
ns_book5[unkown_year].head()

ns_book5.loc[unkown_year, '발행년도'] = '-1'
ns_book5 = ns_book5.astype({'발행년도': 'int32'})

ns_book5['발행년도'].gt(4000).sum()

dangun_yy_rows = ns_book5['발행년도'].gt(4000)
ns_book5.loc[dangun_yy_rows, '발행년도'] = ns_book5.loc[dangun_yy_rows, '발행년도'] - 2333
dangun_year = ns_book5['발행년도'].gt(4000)
print(dangun_year.sum())
ns_book5[dangun_year].head(2)

ns_book5.loc[dangun_year, '발행년도'] = -1

old_books = ns_book5['발행년도'].gt(0) & ns_book5['발행년도'].lt(1900)
ns_book5[old_books]

ns_book5.loc[old_books, '발행년도'] = -1
ns_book5['발행년도'].eq(-1).sum()

# 누락된 정보 채우기

na_rows = ns_book5['도서명'].isna() | ns_book5['저자'].isna() | ns_book5['출판사'].isna() | ns_book5['발행년도'].eq(-1)
print(na_rows.sum())
ns_book5[na_rows].head(2)

import requests
from bs4 import BeautifulSoup

def get_book_title(isbn):
    url = 'http://www.yes24.com/Product/Search?domain=BOOK&query={}'
    r = requests.get(url.format(isbn))
    soup = BeautifulSoup(r.text, 'html.parser')
    title = soup.find('a', attrs={'class':'gd_name'}).get_text()
    return title

get_book_title(9791191266054)

import re

def get_book_info(row):
    title = row['도서명']
    author = row['저자']
    pub = row['출판사']
    year = row['발행년도']
    
    url = 'http://www.yes24.com/Product/Search?domain=BOOK&query={}'
    
    r = requests.get(url.format(row['ISBN']))
    soup = BeautifulSoup(r.text, 'html.parser')
    try:
        if pd.isna(title):
            title = soup.find('a', attrs={'class':'gd_name'}).get_text()
    except AttributeError:
        pass
    try:
        if pd.isna(author):
            aythor = soup.find('span', attrs={'class':'info_auth'}).find_all('a')
            author_list = [auth.get_text() for auth in authors]
            suthor = ','.join(author_list)
    except AttributeError:
        pass
    
    try:
        if pd.isna(pub):
            pub = soup.find('span', attrs = {'class':'info_pub'}).find('a').get_text()
    except AttributeError:
        pass
    
    try:
        if year == -1:
            year_str = soup.find('span', attrss={'class':'info_date'}).get_text()
            year = re.findall(r'\d{4}', year_str)[0]
    except AttributeError:
        pass
    
    return title, author, pub, year
    

updated_sample = ns_book5[na_rows].head(2).apply(get_book_info, axis = 1, result_type='expand')
updated_sample

gdown.download('https://bit.ly/3UJZiHw', 'ns_book5_update.csv', quiet=False)
ns_book5_update = pd.read_csv('ns_book5_update.csv', index_col = 0)
ns_book5_update.head()

ns_book5.update(ns_book5_update)
na_rows = ns_book5['도서명'].isna() | ns_book5['저자'].isna() | ns_book5['출판사'].isna() | ns_book5['발행년도'].eq(-1)
print(na_rows.sum())

ns_book5 = ns_book5.astype({'발행년도':'int32'})
ns_book6 = ns_book5.dropna(subset=['도서명','저자','출판사'])
ns_book6 = ns_book6[ns_book6['발행년도'] != -1]
ns_book6.head()

ns_book6.to_csv('ns_book6.csv',index=False)

# 데이터를 이해하고 올바르게 정제하기

def data_fixing(ns_book4):
    
    ns_book4 = ns_book4.astype({'도서권수':'int32', '대출건수':'int32'})
    set_isbn_na_rows = ns_book4['세트 ISBN'].isna()
    ns_book4.loc[set_isbn_na_rows, '세트 ISBN'] = ''
    ns_book5 = ns_book4.replace({'발행년도':'.*(\d{4}).*'}, r'\1', regex=True)
    unkown_year = ns_book5['발행년도'].str.contains('\D', na=True)
    ns_book5.loc[unkown_year, '발행년도'] = '-1'
    ns_book5.ns_book5.astype({'발행년도': 'int32'})
    
    dangun_yy_rows = ns_book5['발행년도'].gt(4000)
    ns_book.loc[dangun_yy_rows, '발행년도'] = ns_book5.loc[dangun_yy_rows, '발행년도'] - 2333
    
    dangun_year = ns_book5['발행년도'].gt(4000)
    ns_book5.loc[dangun_year, '발행년도'] = -1
    old_books = ns_book5['발행년도'].gt(0) & ns_book5['발행년도'].lt(1900)
    ns_book5.loc[old_books, '발행년도'] = -1
    
    na_rows = ns_book5['도서명'].isna() | ns_book5['저자'].isna() | ns_book5['출판사'].isna() | ns_book5['발행년도'].eq(-1)
    
    updated_sample = ns_book5[na_rows].apply(get_book_info, axis=1, result_type = 'expand')
    
    ns_book6 = ns_book5.dropna(subset=['도서명','저자','출판사'])
    ns_book6 = ns_book6[ns_book6['발행년도'] != -1]
    
    return ns_book6
