==> 4-1.ipynb <==
# 기술통계 구하기

import gdown
gdown.download('https://bit.ly/3736JW1', 'ns_book6.csv', quiet=False)

import pandas as pd
ns_book6 = pd.read_csv('ns_book6.csv',low_memory=False)
ns_book6.head()

ns_book6.describe()

sum(ns_book6['도서권수'] == 0)

ns_book7 = ns_book6[ns_book6['도서권수']>0]

ns_book7.describe(percentiles=[0.3,0.6,0.9])

ns_book7.describe(include='object')

# 평균 구하기

x = [10,20,30]
sum = 0
for i in range(3):
    sum += x[i]
print('평균:', sum/len(x))

ns_book7['대출건수'].mean()

ns_book7['대출건수'].median()

temp_df = pd.DataFrame([1,2,3,4])
temp_df.median()

ns_book7['대출건수'].drop_duplicates().median()

ns_book7['대출건수'].min()

ns_book7['대출건수'].max()

# 분위수 구하기

ns_book7['대출건수'].quantile(0.25)

ns_book7['대출건수'].quantile([0.25,0.5,0.75])

pd.Series([1,2,3,4,5]).quantile(0.9)

borrow_10_flag = ns_book7['대출건수'] < 10

borrow_10_flag.mean()

ns_book7['대출건수'].quantile(0.65)

# 분산 구하기

ns_book7['대출건수'].var()

# 표준편차 구하기

ns_book7['대출건수'].std()

# 최빈값 구하기

ns_book7['도서명'].mode()

ns_book7['발행년도'].mode()

# 데이터프레임에서 기술통계 구하기

ns_book7.mean(numeric_only=True)

ns_book7.loc[:, '도서명':].mode()

ns_book7.to_csv('ns_book7.csv',index=False)

# numpy 기술통계 함수

import numpy as np
np.mean(ns_book7['대출건수'])

np.average(ns_book7['대출건수'], weights=1/ns_book7['도서권수'])

np.mean(ns_book7['대출건수']/ns_book7['도서권수'])

ns_book7['대출건수'].sum()/ns_book7['도서권수'].sum()

np.median(ns_book7['대출건수'])

np.min(ns_book7['대출건수'])

np.max(ns_book7['대출건수'])

np.quantile(ns_book7['대출건수'], [0.25,0.5,0.75])

np.var(ns_book7['대출건수'])

ns_book7['대출건수'].var()

ns_book7['대출건수'].var(ddof=0)

np.var(ns_book7['대출건수'],ddof=1)

np.std(ns_book7['대출건수'])

values, counts = np.unique(ns_book7['도서명'], return_counts=True)

max_idx = np.argmax(counts)

values[max_idx]
